+++ R/zzz.R
@@ -98,7 +98,7 @@ replace_private_with_pub_methods(Series, "^Series_")
 
 
 
-# expression constructors
+# expression constructors, why not just pl$lit = Expr_lit?
 move_env_elements(Expr, pl, c("lit"), remove = FALSE)