+++ src/rust/src/lazy/dataframe.rs
@@ -7,6 +7,7 @@ use crate::rdatatype::new_quantile_interpolation_option;
 use crate::rdatatype::new_unique_keep_strategy;
 use crate::rdatatype::{new_asof_strategy, RPolarsDataType};
 use crate::robj_to;
+use crate::rpolarserr::rerr;
 use crate::rpolarserr::RResult;
 use crate::rpolarserr::{Rctx, WithRctx};
 use crate::utils::wrappers::null_to_opt;
@@ -231,6 +232,8 @@ impl LazyFrame {
     }
 
     fn with_column(&self, expr: &Expr) -> LazyFrame {
+        R!("warning('`with_column()` is deprecated and will be removed in polars 0.9.0. Please use `with_columns()` instead.')")
+        .expect("warning will not fail");
         LazyFrame(self.0.clone().with_column(expr.0.clone()))
     }
 
@@ -390,6 +393,7 @@ impl LazyFrame {
         ))
     }
 
+
     fn fetch(&self, n_rows: Robj) -> RResult<DataFrame> {
         Ok(self
             .0
@@ -397,6 +401,21 @@ impl LazyFrame {
             .fetch(robj_to!(usize, n_rows)?)
             .map_err(crate::rpolarserr::polars_to_rpolars_err)?
             .into())
+
+    fn explode(&self, columns: Robj, dotdotdot_args: Robj) -> RResult<LazyFrame> {
+        let mut columns: Vec<pl::Expr> = robj_to!(Vec, PLExprCol, columns)?;
+        let mut ddd_args: Vec<pl::Expr> = robj_to!(Vec, PLExprCol, dotdotdot_args)?;
+        columns.append(&mut ddd_args);
+        if columns.is_empty() {
+            rerr()
+                .plain("neither have any elements, cannot use explode without Expr(s)")
+                .when("joining Exprs from input [columns] and input [...]")?;
+        }
+        Ok(self.0.clone().explode(columns).into())
+    }
+
+    pub fn clone_see_me_macro(&self) -> LazyFrame {
+        self.clone()
     }
 }